Added file verification method. Improved status display. code cleanup and optimisation

+++ FileFlux/Services/HttpDownloadService.cs
@@ -2,6 +2,7 @@
 using FileFlux.Model;
 
 using System.Net.Http.Headers;
+using System.Security.Cryptography;
 
 namespace FileFlux.Services
 {
@@ -81,50 +82,59 @@ namespace FileFlux.Services
 
                 response.EnsureSuccessStatusCode();
 
-                using var contentStream = await response.Content.ReadAsStreamAsync();
-
-                using var fileStream = new FileStream(fileDownload.SavePath, fileMode, FileAccess.Write, FileShare.None);
-
-
-                var buffer = new byte[8192];
-                
-                int bytesRead;
-
-                while ((bytesRead = await contentStream.ReadAsync(buffer)) > 0)
+                if(response.Headers.ETag != null && response.Headers.ETag.Tag != fileDownload.ETag)
                 {
-
-                    await fileStream.WriteAsync(buffer.AsMemory(0, bytesRead));
-                    totalRead += bytesRead;
-
-                    fileDownload.TotalDownloaded = totalRead;
-                    fileDownload.PercentCompleted = (double)totalRead / fileDownload.TotalSize * 100;
-                    if (fileDownload.CancellationTokenSource.IsCancellationRequested)
-                    {
-                        fileStream.Close();
-                        switch (fileDownload.Status)
-                        {
-                            case FileDownloadStatuses.Paused:
-                                return;
-                            case FileDownloadStatuses.Cancelled:
-                                File.Delete(fileDownload.SavePath);
-                                break;
-                        }
-
-                    }
+                    fileDownload.Status = FileDownloadStatuses.Failed;
+                    fileDownload.ErrorMessage = "The content on the server has changed since the initial download attempt (ETag mismatch).";
+                    return;
                 }
 
+                using var contentStream = await response.Content.ReadAsStreamAsync();
+
+                using (var fileStream = new FileStream(fileDownload.SavePath, fileMode, FileAccess.Write, FileShare.None))
+                {
+                    var buffer = new byte[8192];
+
+                    int bytesRead;
+
+                    while ((bytesRead = await contentStream.ReadAsync(buffer)) > 0)
+                    {
+
+                        await fileStream.WriteAsync(buffer.AsMemory(0, bytesRead));
+                        totalRead += bytesRead;
+
+                        fileDownload.TotalDownloaded = totalRead;
+                        fileDownload.PercentCompleted = (double)totalRead / fileDownload.TotalSize * 100;
+                        if (fileDownload.CancellationTokenSource.IsCancellationRequested)
+                        {
+                            fileStream.Close();
+                            switch (fileDownload.Status)
+                            {
+                                case FileDownloadStatuses.Paused:
+                                    return;
+                                case FileDownloadStatuses.Cancelled:
+                                    File.Delete(fileDownload.SavePath);
+                                    break;
+                            }
+
+                        }
+                    }
+
+                }
+                
                 fileDownload.Status = FileDownloadStatuses.Completed;
 
                 File.SetCreationTime(fileDownload.SavePath, fileDownload.Created);
                 File.SetLastWriteTime(fileDownload.SavePath, fileDownload.LastModified);
                 File.SetLastAccessTime(fileDownload.SavePath, DateTime.Now);
+
             }
             catch (Exception ex)
             {
                 fileDownload.Status = FileDownloadStatuses.Failed;
                 fileDownload.ErrorMessage = ex.Message;
             }
-        }
+        }        
 
         public async Task PauseDownload(FileDownload fileDownload)
         {
